requirements.txt
# Optimized Core AI and ML Dependencies
openai>=1.0.0,<2.0.0
langchain>=0.3.0,<0.4.0
langchain-openai>=0.3.0,<0.4.0
tiktoken>=0.11.0,<1.0.0

# Optimized Azure SDK with version pinning
azure-identity>=1.25.0,<2.0.0
azure-mgmt-resource>=24.0.0,<25.0.0
azure-mgmt-compute>=36.0.0,<37.0.0
azure-mgmt-network>=29.0.0,<30.0.0
azure-mgmt-storage>=23.0.0,<24.0.0
azure-cosmos>=4.9.0,<5.0.0
azure-keyvault-secrets>=4.10.0,<5.0.0
azure-monitor-query>=2.0.0,<3.0.0
azure-storage-blob>=12.26.0,<13.0.0

# Additional Azure services
azure-mgmt-costmanagement>=4.0.0
azure-mgmt-consumption>=10.0.0
azure-mgmt-advisor>=9.0.0
azure-mgmt-security>=7.0.0
azure-mgmt-policyinsights>=1.1.0

# Optimized Web Framework
fastapi>=0.116.0,<1.0.0
uvicorn[standard]>=0.35.0,<1.0.0
pydantic>=2.11.0,<3.0.0
pydantic-settings>=2.10.0,<3.0.0
python-multipart>=0.0.20,<1.0.0

# Optimized Database and Cache
redis>=5.0.0,<6.0.0
sqlalchemy>=2.0.0,<3.0.0
asyncpg>=0.29.0  # High-performance PostgreSQL driver
aiosqlite>=0.20.0  # SQLite async support

# Core ML and Data Science
scikit-learn>=1.4.0
pandas>=2.2.0
numpy>=1.26.0
scipy>=1.12.0
joblib>=1.3.0

# Async and Background Tasks
celery>=5.3.0
motor>=3.3.0
aiohttp>=3.9.0
aiofiles>=23.2.0

# Authentication and Security
msal>=1.24.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
cryptography>=42.0.0

# Monitoring and Logging
prometheus-client>=0.19.0
sentry-sdk>=1.39.0

# Optimized Utilities
python-dotenv>=1.0.0,<2.0.0
pyyaml>=6.0.0,<7.0.0
click>=8.1.0,<9.0.0
rich>=13.7.0,<14.0.0
httpx>=0.26.0,<1.0.0
orjson>=3.9.0  # Fast JSON serialization
typer>=0.12.0  # Modern CLI framework

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Configuration Management
watchdog>=3.0.0
alembic>=1.13.0
psycopg2-binary>=2.9.0